refactor: 주석 제거 및 점수 계산 테스트 수정
